refactor(distributions): consolidate duplicated reshape helpers
Move the byte-identical `_axis_names`, `_reshape_batch`, and
`_reshape_samples` helpers (previously copy-pasted across `_mvn.py`,
`_mvn_prec.py`, and `_gp/_kronecker_gp.py`) into a shared
`_distributions/_utils.py` module.

Switch the reshape helpers from einops `rearrange` to `einx.id` and add
einx as a dependency. A repo-wide migration from einops to einx is
tracked separately.
