+++ src/gaussx/_distributions/_mvn.py
@@ -14,53 +14,11 @@ from gaussx._distributions._gaussian import (
     _gaussian_log_prob_residual,
     gaussian_entropy,
 )
+from gaussx._distributions._utils import _reshape_batch, _reshape_samples
 from gaussx._primitives._cholesky import cholesky as _cholesky
 from gaussx._primitives._diag import diag as _diag
 from gaussx._strategies._auto import AutoSolver
 from gaussx._strategies._base import AbstractSolverStrategy
-
-
-def _axis_names(count: int) -> tuple[str, ...]:
-    names = []
-    for index in range(count):
-        value = index
-        chars = []
-        while True:
-            value, remainder = divmod(value, 26)
-            chars.append(chr(ord("a") + remainder))
-            if value == 0:
-                break
-            value -= 1
-        names.append("".join(reversed(chars)))
-    return tuple(names)
-
-
-def _reshape_batch(
-    values: Float[Array, " flat"],
-    batch_shape: tuple[int, ...],
-) -> Float[Array, "*batch"]:
-    if not batch_shape:
-        return values[0]
-    batch_axes = _axis_names(len(batch_shape))
-    axis_lengths = dict(zip(batch_axes, batch_shape, strict=True))
-    batch_pattern = " ".join(batch_axes)
-    return rearrange(values, f"({batch_pattern}) -> {batch_pattern}", **axis_lengths)
-
-
-def _reshape_samples(
-    values: Float[Array, "flat N"],
-    batch_shape: tuple[int, ...],
-) -> Float[Array, "*batch N"]:
-    if not batch_shape:
-        return values[0]
-    batch_axes = _axis_names(len(batch_shape))
-    axis_lengths = dict(zip(batch_axes, batch_shape, strict=True))
-    batch_pattern = " ".join(batch_axes)
-    return rearrange(
-        values,
-        f"({batch_pattern}) N -> {batch_pattern} N",
-        **axis_lengths,
-    )
 
 
 class MultivariateNormal(dist.Distribution):
